Neo to Memgraph importer tool sketch added
Summary:
This importer is Bolt based. It loads data from Neo and directly injects into
Memgraph. Persistence is optional (JSON based). Currently it works for small
datasets, to make it work for larger datasets batching is required. For it to
work more efficiently indexing on label+property is required (in progress).

An in-Memgraph implementation of JSON importing could be much more efficient,
and is compatible with this script. This sript could be used for exporting
from Neo, and a C++ importer tool for the Memgraph side import.

I suggest that Gradicek takes this over and I can move on to other tasks
I have (if Gradicek is available for this).

Reviewers: buda, matej.gradicek

Reviewed By: matej.gradicek

Subscribers: pullbot

Differential Revision: https://phabricator.memgraph.io/D524