Remove rechecking external projects
Summary:
This is probably a bug in CMake. When the external projects are forced
to be rechecked, CPack takes their outputs and puts them in binary
release.

Since rechecking is removed, this means that any changes to external
projects (like gflags or glog) will not be recompiled with plain
`make all`. You need to call `make clean` first. The same applies to
catching configuration like removal of filenames in glog when creating a
release build.

Reviewers: mferencevic

Reviewed By: mferencevic

Subscribers: pullbot

Differential Revision: https://phabricator.memgraph.io/D890